+++ server/src/main/kotlin/com/heerkirov/hedge/server/dto/form/SourceData.kt
@@ -7,11 +7,11 @@ data class SourceDataBulkForm(val items: List<SourceDataCreateForm>)
 
 data class SourceDataCreateForm(val sourceSite: String, val sourceId: Long, val status: Opt<SourceEditStatus>,
                                 val title: Opt<String?>, val description: Opt<String?>,
-                                val tags: Opt<List<SourceTagForm>>, val books: Opt<List<SourceBookForm>>, val relations: Opt<List<Int>>
+                                val tags: Opt<List<SourceTagForm>>, val books: Opt<List<SourceBookForm>>, val relations: Opt<List<Long>>
 )
 
 data class SourceDataUpdateForm(val title: Opt<String?>, val description: Opt<String?>, val status: Opt<SourceEditStatus>,
-                                val tags: Opt<List<SourceTagForm>>, val books: Opt<List<SourceBookForm>>, val relations: Opt<List<Int>>)
+                                val tags: Opt<List<SourceTagForm>>, val books: Opt<List<SourceBookForm>>, val relations: Opt<List<Long>>)
 
 data class SourceTagForm(val code: String, val name: Opt<String>, val otherName: Opt<String?>, val type: Opt<String?>)
