+++ server/src/main/kotlin/com/heerkirov/hedge/server/dto/res/SourceData.kt
@@ -11,7 +11,7 @@ data class SourceDataRes(val sourceSite: String, val sourceSiteName: String, val
 data class SourceDataDetailRes(val sourceSite: String, val sourceSiteName: String, val sourceId: Long,
                                val title: String, val description: String,
                                val empty: Boolean, val status: SourceEditStatus,
-                               val tags: List<SourceTagDto>, val books: List<SourceBookDto>, val relations: List<Int>,
+                               val tags: List<SourceTagDto>, val books: List<SourceBookDto>, val relations: List<Long>,
                                val createTime: LocalDateTime, val updateTime: LocalDateTime)
 
 data class SourceTagDto(val code: String, val name: String, val otherName: String?, val type: String?)
